Updated color palette and added limits for ethnicity.
Courtney mentioned that the limit changed for the ethnicity plots when you clicked on certain filters, so I added the 'limits' parameter to the ethnicity plot. One question that raises is that of repeated colors: because there are so many categories for ethnicity, a couple colors are repeated, and I had to add a couple colors that were not part of the original colorblind friendly palette.